requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch
torchvision
numpy
scipy
scikit-learn
numpy  
matplotlib 
ipython 
pandas 
sympy 
nose
h5py
Pillow
opencv-python